% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm.R
\name{ccm}
\alias{ccm}
\title{ccm}
\usage{
ccm(ts, model = 1)
}
\arguments{
\item{x}{numeric vector of time series data}

\item{title}{graph title}
}
\description{
ccm() estimated convergent cross mapping prediciton skill
}
\details{
ccm() is from HBR, Code 8.1, modified by BPR
}
\references{
Huffaker, Ray; Bittelli, Marco; Rosa, Rodolfo. Nonlinear Time Series Analysis with R (p. 224). (Function). Kindle Edition.
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{cobweb}
\keyword{plot}
