% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobweb.R
\name{cobweb}
\alias{cobweb}
\title{cobweb}
\usage{
cobweb(x, title = NULL)
}
\arguments{
\item{x}{numeric vector of time series data}

\item{title}{graph title}
}
\description{
cobweb() creates a cobweb plot of the data.
}
\details{
cobweb() creates a cobweb plot of the data. It plots the consecutive points
of the data, connects adjacent points with lines, and draws a diagonal line
on the graph.
}
\examples{
3.99999 -> R
vector(mode = "numeric", length = 500) -> x
0.7 -> x[1]
for(i in 2:500) {R * x[i-1] * (1 - x[i-1]) -> x[i]}
cobweb(x, paste0("R = ",R))

}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{cobweb}
\keyword{plot}
