% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const_delay_embed.R
\name{const_delay_embed}
\alias{const_delay_embed}
\title{const_delay_embed}
\usage{
const_delay_embed(x, m.max = 6, max.delay = 20, lag.max = 100, tol = 0.15)
}
\arguments{
\item{x}{numeric vector (representing time series)}

\item{m.max}{maximum embedding dimension to consider}

\item{max.delay}{maximum delay to consider}

\item{lag.max}{maximum lag to consider for autocorrelation}

\item{tol}{Default tolerance for "closeness" of false nearest neighbors}
}
\description{
const_delay_embed() determines a constant delay reconstruction state space
and embeds the time series in it.
}
\details{
const_delay_embed() estimates appropriate values for delay, embedding
dimension, and Theiler window, and then uses a false-nearest neighbors
approach to determine the appropriate embedding of a time series into a
reconstructed state space.
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{delay,}
\keyword{dimension}
\keyword{embedding}
\keyword{embedding,}
