% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMBestimates.R
\name{EMBestimates}
\alias{EMBestimates}
\title{Create a Empirical Bayes estimate of derivatives}
\usage{
EMBestimates(DataM, Nobs, maxorder = 2, dt = NA, TimeM = NA)
}
\arguments{
\item{DataM}{A data sequence}

\item{Nobs}{Embedding dimension}

\item{maxorder}{The highest order of derivative to estimate}

\item{dt}{NA means all time intervals are equal}

\item{TimeM}{Times of the data (not used of dt is NA)}
}
\value{
A matrix of derivatives, in columns
}
\description{
Create a Empirical Bayes estimate of derivatives
}
\examples{
set.seed(42)
1:100 -> t
4 * exp(-t/20) * sin(pi * t / 5) + rnorm(100, 0, 0.1) -> x
EMBestimates(x, 2, 2) -> dx
}
