% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_burstiness.R
\name{event_burstiness}
\alias{event_burstiness}
\title{event_burstiness}
\usage{
event_burstiness(ts, min_iet = NULL)
}
\arguments{
\item{<ts>}{character (or logical) vector}

\item{<min_iet>}{minimum interevent sequence spacing}
}
\description{
event_burstiness() estimates the burstiness coefficient for
each of the unique codes in a categorical time series. The series is
considered to be event-based (i.e., the points are ordered in time but the
intervals between points may not be constant). See time_burstiness() to find
the burstiness of codes in a time-series.

There is a known problem with a non-NULL minimum inter-event time. That
needs to be corrected.
}
\details{
event_burstiness() returns a named vector with a burstiness number for
each unique code in the data sequence, including blank or missing codes.

Standard errors for the resulting burstiness coefficients can be
obtained using block resampling (e.g., via boot::tsboot() with the
parameter sim = "fixed").
}
\examples{
event_burstiness(guastello, 1)
event_burstiness(guastello)
event_burstiness(engineering)

}
\references{
Kim, E.-K., & Jo, H.-H. (2016). Measuring burstiness for finite
event sequences. Physical Review E, 94(3), 032311.
https://doi.org/10.1103/PhysRevE.94.032311
}
\seealso{
tsboot()
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{burstiness}
