% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_1d.R
\name{fixed_1d}
\alias{fixed_1d}
\title{Estimate the fixed points of a 1-dimensional nonlinear system.}
\usage{
fixed_1d(Level_data, Velocity_data, span = 0.3, numPred = 100)
}
\arguments{
\item{Level_data}{A vector of the level (0th derivative) data.}

\item{Velocity_data}{A vector of the velocity (1st derivative) data.}

\item{span}{Smoothing for the loess() function. (Default is 0.3.)}

\item{numPred}{Size of grid on which to predict velocity as a function of
level. (Default is 100.)}
}
\value{
A data frame with columns "Level" indicating the estimated fixed
    point and "Stability" indicating whether the fixed point is stable of
    unstable.
}
\description{
This function estimates the zeroes in the function velocity(level).
}
