% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_field.R
\name{gradient_field}
\alias{gradient_field}
\title{Plot a vector field and a data path through that field}
\usage{
gradient_field(
  df,
  xvar,
  yvar,
  gam_form,
  inter = TRUE,
  xdot = NULL,
  ydot = NULL,
  spar = 0.6,
  arrow_scale = 10,
  grid_size = 50,
  display_grid_size = 20
)
}
\arguments{
\item{df}{A data frame of data to fit. The first four columns must be the
(x,y) coordinates and the x- and y-directional derivatives, respectively.}

\item{arrow_scale}{A scaling factor for the arrow lengths. Larger scales
produce shorter arrows. This will be changed at some point.}

\item{data_df}{A data frame of the path (x,y) coordinates in temporal order.}

\item{arrow_size}{The base size of the arrow heads, in centimeters.}
}
\value{
A ggplot2 object with the vector field plot and path. Labels, main
    title, etc., may be added after the return.
}
\description{
Plot a vector field and a data path through that field
}
