% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPL_fit.R
\name{IPL_fit}
\alias{IPL_fit}
\title{IPL_fit}
\usage{
IPL_fit(x, estimator = "CSN", gof = "logLik")
}
\arguments{
\item{x}{numeric or character}

\item{estimator}{Estimation method ("CSN", "Hill", "moments" or "peaks", or "Pickands")}

\item{gof}{Goodness of fit measure (pseudo-R-squared, KS, or loglikelihood)}
}
\description{
This function estimates the (inverse) power-law fit to a
dataset using one of several estimators.
}
\details{
IPL_fit returns the fit parameters and a goodness of fit measure for a
power-law fit to data.
}
\references{
Nair, J., Wierman, A., & Zwart, B. (2022). The Fundamentals of
 heavy tails: Properties, emergence, and estimation. Cambridge University
 Press. (See Chapter 9.)
}
\author{
Barney Ricca <barneyricca@gmail.com>
}
\keyword{power-law}
