% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_ortho_embed.R
\name{max_ortho_embed}
\alias{max_ortho_embed}
\title{max_ortho_embed}
\usage{
max_ortho_embed(ts, max_lag, max_dim, method = "mutual")
}
\arguments{
\item{ts}{numeric vector of the time series to evaluate}

\item{max_lag}{maximum lag to consider}

\item{max_dim}{maximum dimension to allow}

\item{method}{method choice; not yet implemented so do not use.}
}
\description{
This function computes the delay state space embedding of a
time series, based on the method proposed by Pecora et al. (2008).
}
\details{
max_ortho_embed returns a (possibly non-constant) delay state space
embedding of ts. It does this by determining the maximally independent
delay vector at each step. The maximum embedding dimenions for the naive
approach (the only one implemented so far) is 6. Eventually, this will
include 3 methods; the mutual information method implemented here, the
method proposed by Pecora et al., and an approximation of the Pecora method.
The naive approach (method = "mutual") is a slight extension of the constant
delay embedding approach.
}
\references{
Pecora, L. M., Moniz, L., Nichols, J., & Carroll, T. L. (2007).
A unified approach to attractor reconstruction. Chaos, 17(1),
013110-1-013110–013119. https://doi.org/10.1063/1.2430294
}
\author{
Barney Ricca <barneyricca@gmail.com>
}
\keyword{embedding}
