% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayport.R
\docType{data}
\name{mayport}
\alias{mayport}
\title{mayport}
\format{
a data frame with 2 columns (Time and Level)
}
\usage{
mayport
}
\description{
Sea level data from the Mayport station.
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{level,}
\keyword{mayport}
\keyword{sea}
