% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse}
\alias{nse}
\title{nse}
\usage{
nse(ts, frac.learn = 0.5)
}
\arguments{
\item{<ts>}{(1-dimensional) numeric vector or (2-dimensional) matrix. If a
vector, then it is assumed ts is a data stream to be embedded in a shadow
state space. If a matrix, it is assumed that ts represents a data stream
embedded in a reconstructed state space.}

\item{<frac.learn>}{fraction of data in learning set}
}
\description{
nse() estimates the Nash Sutcliffe efficiency for a time series.
}
\details{
nse() returns a list which includes the Nash-Sutcliffe efficiency.
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{Nash}
\keyword{Sutcliffe}
\keyword{efficiency}
