% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_fit.R
\name{ode_fit}
\alias{ode_fit}
\title{Estimate unknown coefficients of a first order differential equation}
\usage{
ode_fit(
  df,
  user_func,
  parm_init,
  parm_names,
  init_step_size = 0.05,
  max_iter = 1000,
  rtol = 0.001,
  tol = 0.1
)
}
\arguments{
\item{df}{A data frame of data to fit. The first column must contain the
times.}

\item{user_func}{The ODE function. Must correspond to the requirements
of the deSolve::ode() function.}

\item{parm_init}{Initial coefficient estimates}

\item{parm_names}{The parameter names. These must match those used in
user_func, but do not need to match the column names of df.}

\item{init_step_size}{Initial step size for each iteration}

\item{max_iter}{The maximum number of steps to take.}

\item{rtol}{The smallest step size}

\item{tol}{Smallest RMSE improvement to continue iterating}
}
\value{
A list of estimated parameters and process information.
}
\description{
Estimate unknown coefficients of a first order differential equation
}
