% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbital_decomposition.R
\name{orbde}
\alias{orbde}
\title{orbde}
\usage{
orbde(data_seq)
}
\arguments{
\item{<data_seq>}{character or integer time series}
}
\description{
This function returns the result of orbital decomposition.
All of the metrics associated with orbital decomposition are returned in a
matrix with named columns.
}
\examples{
orbde(guastello)

}
\references{
Guastello, S. J. (n.d.). Orbital Decomposition: Identification
of Dynamical Patterns in Categorical Data. In S. J. Guastello & R. A. M.
Gregson (Eds.), Nonlinear Dynamical Systems Analysis for the Behavioral
Sciences Using Real Data (pp. 499–516). CRC Press.

Guastello, S. J., Hyde, T., & Odak, M. (1998). Symbolic dynamic patterns of
verbal exchange in a creative problem solving group. Nonlinear Dynamics,
Psychology, and Life Sciences, 2(1), 35–58.
}
\author{
Bernard P. Ricca bricca@uccs.edu
}
\keyword{decomposition}
\keyword{orbital}
