% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdcor.R
\name{pdcor}
\alias{pdcor}
\title{Estimate the partial distance correlation of two vectors.}
\usage{
pdcor(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as x.}
}
\value{
The partial distance correlation of the two vectors x and y.
}
\description{
This function returns the partial distance correlation of two vectors. It
    does so by removing the linear fit between the vectors before estimating
    the distance correlation.
}
