% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdcor_ci.R
\name{pdcor_ci}
\alias{pdcor_ci}
\title{Estimate the partial distance correlation, with CI, between two vectors.}
\usage{
pdcor_ci(x, y, R = 1000)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as x.}

\item{R}{The number of replications. Default is 1000.}
}
\value{
A named vector with the partial distance correlation ($mn) of the
    two vectors x and y, along with the lower ($lCI) and upper ($uCI)
    confidence interval limits.
}
\description{
This function returns the partial distance correlation of two
    vectors along with a bootstrapped 95% confidence interval. It does so by
    removing the linear fit between the vectors before estimating the
    distance correlation.
}
