% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSG.R
\name{SSG}
\alias{SSG}
\title{SSG}
\usage{
SSG(ts1, ts2, times = NULL, cats1 = NULL, cats2 = NULL)
}
\arguments{
\item{<x>}{text file name}
}
\description{
SSG creates a state space grid of the data
}
\details{
This function returns a numeric sequence corresponding to a text
file, suitable for RQA use.


SSG creates and plots a state space grid of the data.
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{clean}
\keyword{text}
