% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSG.R
\name{SSG_metrics}
\alias{SSG_metrics}
\title{SSG_metrics}
\usage{
SSG_metrics(ts1, ts2, times = NULL, cats1 = NULL, cats2 = NULL)
}
\arguments{
\item{<x>}{text file name}
}
\description{
This function returns a numeric sequence corresponding to a text
file, suitable for RQA use.
}
\examples{
SSG_metrics(ts)

}
\keyword{clean}
\keyword{text}
