% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisFile.R
\name{thisFile}
\alias{thisFile}
\title{Path to current file}
\usage{
thisFile()
}
\description{
This function will return the path to the current file it
is called from. Will work either script is run from Rscript
or is sourced. Why the F* is this so hard in R
}
