% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_burstiness.R
\name{time_burstiness}
\alias{time_burstiness}
\title{time_burstiness}
\usage{
time_burstiness(ts, min_iet = NULL)
}
\arguments{
\item{ts}{numeric vector (representing event times).}

\item{min_iet}{minimum interevent sequence spacing}
}
\description{
burstiness() estimates the burstiness coefficient for events.
}
\details{
burstiness() returns a vector with an estimated burstiness coefficient for
the events whose times are given in the data sequence.

There is a known problem with a non-NULL minimum inter-event time. That
needs to be corrected.
}
\examples{
time_burstiness(poisson_times)
time_burstiness(poisson_times, 2)

}
\references{
Kim, E.-K., & Jo, H.-H. (2016). Measuring burstiness for finite event
sequences. Physical Review E, 94(3), 032311.
https://doi.org/10.1103/PhysRevE.94.032311
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{burstiness}
