% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfp_plot.R
\name{vfp_plot}
\alias{vfp_plot}
\title{Plot a vector field and a data path through that field}
\usage{
vfp_plot(df, data_df, arrow_size = 0.2, arrow_scale = 1)
}
\arguments{
\item{df}{A data frame of data to fit. The first four columns must be the
(x,y) coordinates and the x- and y-directional derivatives, respectively.}

\item{data_df}{A data frame of the path (x,y) coordinates in temporal order.}

\item{arrow_size}{The base size of the arrow heads, in centimeters.}

\item{arrow_scale}{A scaling factor for the arrow lengths. Larger scales
produce shorter arrows.}
}
\value{
A ggplot2 object with the vector field plot and path. Labels, main
    title, etc., may be added after the return.
}
\description{
This function takes a data frame of directional derivatives and a data
    frame of data points, and creates a vector field plot with the path
    superposed on it. The current version DOES NOT check for appropriate
    input, so be careful.
}
