% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guastello.R
\docType{data}
\name{guastello}
\alias{guastello}
\title{guastello}
\format{
a character vector of 20 entries
}
\source{
<https://www.routledge.com/Nonlinear-Dynamical-Systems-Analysis-for-the-Behavioral-Sciences-Using-Real-Data/Guastello-Gregson/p/book/9781439819975?srsltid=AfmBOooMo5hYC4OG-x8bUNkPVhBbLo-IyQJ-XCyF-5YG8HDlxT238AvN>
}
\usage{
guastello
}
\description{
This short (20 entries) sequence of the letters "A" through "E"
is taken from Table 21.1 of Guastello's 2011 introduction to orbital
decomposition. It was originally included here to provide test data for
the orbde() function.
}
\examples{
burstiness(guastello, 1)
burstiness(guastello)

}
\references{
Guastello, S. J. (2011). Orbital decomposition: Identification
of dynamical patterns in categorical data. In S. J. Guastello & R. A. M.
Gregson (Eds.), Nonlinear dynamical systems analysis for the behavioral
sciences using real data (pp. 499–516). CRC Press.
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{guastello}
