% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_ortho_embed.R
\name{max_ortho_embed}
\alias{max_ortho_embed}
\title{max_ortho_embed}
\usage{
max_ortho_embed(ts1)
}
\arguments{
\item{ts1}{character or numeric vector; time series to evaluate}
}
\description{
This function computes the delay state space embedding of a
time series, based on the method proposed by Pecora et al. (2008).
}
\details{
It returns
}
\references{
Pecora, L. M., Moniz, L., Nichols, J., & Carroll, T. L. (2007).
A unified approach to attractor reconstruction. Chaos, 17(1),
013110-1-013110–013119. https://doi.org/10.1063/1.2430294
}
\author{
Barney Ricca <barneyricca@gmail.com>
}
\keyword{embedding}
