% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_ksg.R
\name{mi_ksg}
\alias{mi_ksg}
\title{mi_ksg}
\usage{
mi_ksg(ts, k = 3, lag.max = 100)
}
\arguments{
\item{<ts>}{numeric vector of time series data}

\item{<k>}{number of nearest neighbors to use}

\item{<lag.max>}{Maximum lag to use}
}
\description{
mi_ksg() uses the approach of Kraskov, Stögbauer, and Grassberger (KSG) to
estimate the mutual information in a time-series.
}
\details{
mi_ksg() uses the approach of Kraskov, Stögbauer, and Grassberger (KSG) to
estimate the mutual information in a time-series. It returns a vector of
mutual information estimates at each lag up to lag.max(), starting with
a lag of zero. (Hence, the lags are one less than the indices of the
vector.)
}
\references{
Kraskov, A., Stögbauer, H., & Grassberger, P. (2004).
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{information}
\keyword{mutual}
