% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDSWorkshop-package.R
\docType{package}
\name{ndstools-package}
\alias{ndstools}
\alias{ndstools-package}
\title{ndstools: A Collection of Useful Tools for Nonlinear Time Series Analyses}
\description{
This package contains a number of tools for nonlinear time series (NLTS) analyses. Some of these tools are available elsewhere, but were ported to Rcpp here for speed. Some of these tools are modifications and encapsulations of code that can be found in Huffaker, Bitelli, and Rosa (2017). Some of these tools are new.
}
\author{
\strong{Maintainer}: Barney Ricca \email{barneyricca@gmail.com}

}
\keyword{internal}
