% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSG.R, R/SSG_metrics.R
\name{SSG}
\alias{SSG}
\title{SSG}
\usage{
SSG(ts1, ts2, times = NULL, cats1 = NULL, cats2 = NULL)

SSG(ts1, ts2, times = NULL, cats1 = NULL, cats2 = NULL)
}
\arguments{
\item{<x>}{text file name}
}
\description{
This function returns a numeric sequence corresponding to a text
file, suitable for RQA use.

This function returns a numeric sequence corresponding to a text
file, suitable for RQA use.
}
\keyword{clean}
\keyword{text}
