% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_burstiness.R
\name{time_burstiness}
\alias{time_burstiness}
\title{time_burstiness}
\usage{
time_burstiness(times, min_time = 0, min_iet = NULL)
}
\arguments{
\item{times}{numeric vector}

\item{min_time}{the time at which the data sequence begins (assumed to be
0)}

\item{min_iet}{minimum interevent sequence spacing}
}
\description{
time_burstiness() estimates the burstiness coefficient for a
sequence conataining the series of times at which an event occurs.

See burstiness() to find the burstiness of a time-series of categorical
data.

There is a known problem with a non-NULL minimum inter-event time. That
needs to be corrected.
}
\details{
time_burstiness() returns the burstiness number for a sequence of event
times (e.g., the successive times at which an event occurs).

Standard errors for the resulting burstiness coefficients can be
obtained using block resampling, but that has not been implemented here yet.
(It seems that, unlike burstiness(), boot:tsboot() will not return the
correct standard errors for time_burstiness().)

There may still be a bug in the non-NULL minimum inter-event times
parameter.
}
\references{
Kim, E.-K., & Jo, H.-H. (2016). Measuring burstiness for finite
event sequences. Physical Review E, 94(3), 032311.
https://doi.org/10.1103/PhysRevE.94.032311
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{burstiness}
